feat(newsletter): add subcripstion page
