+++ home/urls.py
@@ -21,14 +21,14 @@ urlpatterns = [
         name="nwl-subscription",
     ),
     path(
-        "newsletter/confirmation/<slug:token>",
-        views.NewsLetterConfirmationView.as_view(),
+        "newsletter/confirmation",
+        views.NewsletterConfirmationSubscriptionView.as_view(),
         name="nwl-confirmation",
     ),
     path(
-        "emails",
-        views.AllEmailsView.as_view(),
-        name="all-emails",
+        "newsletter/confirmation/<slug:token>",
+        views.NewsletterEmailValidationView.as_view(),
+        name="nwl-validation",
     ),
     path("maintenance", views.MaintenanceView.as_view(), name="maintenance_mode"),
     path("alive/async-workers", views.AliveView.as_view(), name="alive"),
