+++ home/views/__init__.py
@@ -1,0 +1,35 @@
+from .AccessView import AccessView
+from .AliveView import AliveView
+from .ContactView import ContactView
+from .download_package_request import download_package_request
+from .DownloadView import DownloadView
+from .HomeRapportLocalView import HomeRapportLocalView
+from .HomeView import HomeView
+from .LegalNoticeView import LegalNoticeView
+from .MaintenanceView import MaintenanceView
+from .NewsletterConfirmationSubscriptionView import (
+    NewsletterConfirmationSubscriptionView,
+)
+from .NewsletterEmailValidationView import NewsletterEmailValidationView
+from .NewsletterSubscriptionView import NewsletterSubscriptionView
+from .PrivacyView import PrivacyView
+from .RobotView import RobotView
+from .StatsView import StatsView
+
+__all__ = [
+    "AccessView",
+    "AliveView",
+    "ContactView",
+    "download_package_request",
+    "DownloadView",
+    "HomeRapportLocalView",
+    "HomeView",
+    "LegalNoticeView",
+    "MaintenanceView",
+    "NewsletterSubscriptionView",
+    "NewsletterEmailValidationView",
+    "NewsletterConfirmationSubscriptionView",
+    "PrivacyView",
+    "RobotView",
+    "StatsView",
+]
